kernels, tests: it's doomed
Naive multiplication inherently cannot handle this approach of accumulation, it's inevitable that we will run into overflow. The algorithm breaks down when the inputs are exactly 20 digits long.

At 20 digits, the inputs are represented by 2 full uint32 words. When we multiply, certain uint64 positions accumulate 2 products of full uint32s, resulting in overflow.

I guess we need some kind of intermediate carry propagation, or we can just describe this as a limitation on naive multiplication algorithms and just move over to Schonhage-Strassen.